Split apart test/unit/model/util_test.cpp
