+++ src/stan/model/log_prob_grad.hpp
@@ -95,7 +95,7 @@ namespace stan {
         throw;
       }
     }
-    
+
   }
 }
 #endif
