+++ src/stan/model/log_prob_propto.hpp
@@ -97,7 +97,7 @@ namespace stan {
       stan::math::recover_memory();
       return lp;
     }
-    
+
   }
 }
 #endif
