+++ src/test/unit/model/model_functional_test.cpp
@@ -1,0 +1,3 @@
+#include <stan/model/model_functional.hpp>
+#include <gtest/gtest.h>
+
